// src/components/interfaces/formData.ts
export interface formData {
  name: string;
  email: string;
  company: string;
  role: string;
  phone?: string | undefined;
  file?: object;
  services: string;
  tellUs: string;
}
export interface formWorkData {
  name: string;
  email: string;
  phone: string;
  file: object;
  tellUs: string;
}